--- a/constantine/src/types/kzg_settings.rs
+++ b/constantine/src/types/kzg_settings.rs
@@ -101,8 +101,11 @@ impl KZGSettings<CtFr, CtG1, CtG2, CtFFTSettings, CtPoly, CtFp, CtG1Affine> for 
             g1_values_lagrange_brp: g1_lagrange_brp.to_vec(),
             g2_values_monomial: g2_monomial.to_vec(),
             fs: fft_settings.clone(),
+            precomputation: precompute(g1_lagrange_brp, &x_ext_fft_columns)
+                .ok()
+                .flatten()
+                .map(Arc::new),
             x_ext_fft_columns,
-            precomputation: precompute(g1_lagrange_brp).ok().flatten().map(Arc::new),
             cell_size,
         })
     }
@@ -271,8 +274,8 @@ impl KZGSettings<CtFr, CtG1, CtG2, CtFFTSettings, CtPoly, CtFp, CtG1Affine> for 
         self.precomputation.as_ref().map(|v| v.as_ref())
     }
 
-    fn get_x_ext_fft_column(&self, index: usize) -> &[CtG1] {
-        &self.x_ext_fft_columns[index]
+    fn get_x_ext_fft_columns(&self) -> &[Vec<CtG1>] {
+        &self.x_ext_fft_columns
     }
 
     fn get_cell_size(&self) -> usize {