--- a/kzg/Cargo.toml
+++ b/kzg/Cargo.toml
@@ -9,6 +9,7 @@ num_cpus = { version = "1.16.0", optional = true }
 rayon = { version = "1.8.0", optional = true } 
 threadpool = { version = "^1.8.1", optional = true }
 siphasher = { version = "1.0.0", default-features = false }
+hashbrown = "0.15.2"
 
 [features]
 default = [