--- a/mcl/src/types/kzg_settings.rs
+++ b/mcl/src/types/kzg_settings.rs
@@ -103,8 +103,13 @@ impl KZGSettings<MclFr, MclG1, MclG2, MclFFTSettings, MclPoly, MclFp, MclG1Affin
             g1_values_lagrange_brp: g1_lagrange_brp.to_vec(),
             g2_values_monomial: g2_monomial.to_vec(),
             fs: fft_settings.clone(),
+            precomputation: {
+                precompute(g1_lagrange_brp, &x_ext_fft_columns)
+                    .ok()
+                    .flatten()
+                    .map(Arc::new)
+            },
             x_ext_fft_columns,
-            precomputation: { precompute(g1_lagrange_brp).ok().flatten().map(Arc::new) },
             cell_size,
         })
     }
@@ -274,8 +279,8 @@ impl KZGSettings<MclFr, MclG1, MclG2, MclFFTSettings, MclPoly, MclFp, MclG1Affin
         self.precomputation.as_ref().map(|v| v.as_ref())
     }
 
-    fn get_x_ext_fft_column(&self, index: usize) -> &[MclG1] {
-        &self.x_ext_fft_columns[index]
+    fn get_x_ext_fft_columns(&self) -> &[Vec<MclG1>] {
+        &self.x_ext_fft_columns
     }
 
     fn get_cell_size(&self) -> usize {