--- a/run-c-kzg-4844-benches.sh
+++ b/run-c-kzg-4844-benches.sh
@@ -65,13 +65,20 @@ cargo bench
 
 ###################### go benchmarks ######################
 
-print_msg "Patching go binding"
-git apply < ../go.patch
-cd bindings/go || exit
+if [ "$backend" != "constantine" ]; then
+  print_msg "Patching go binding"
+  git apply < ../go.patch
+  cd bindings/go || exit
 
-print_msg "Running go benchmarks"
-CGO_CFLAGS="-O2 -D__BLST_PORTABLE__" go test -run ^$ -bench .
-cd ../../..
+  print_msg "Running go benchmarks"
+  CGO_CFLAGS="-O2 -D__BLST_PORTABLE__" go test -run ^$ -bench .
+  cd ../..
+else
+  # TODO: fix this, constantine was working through go bindings previously
+  print_msg "Currently, go bindings are not supported for constantine backend"
+fi
+
+cd ..
 
 ###################### cleaning up ######################