@ArtiomTr 2025-02-12
Yeah, I've replaced std HashMap & HashSet with [`hashbrown`](https://github.com/rust-lang/hashbrown) because of no_std support, so I want to re-run benchmarks to make sure that performance haven't degraded